// data/github/current/ui/packages/item-picker/components/__generated__/MilestonePickerSearchQuery.graphql.ts
export type MilestonePickerSearchQuery$variables = {
  count: number;
  owner: string;
  query?: string | null | undefined;
  repo: string;
};
export type MilestonePickerSearchQuery$data = {
  readonly repository: {
    readonly milestones: {
      readonly nodes: ReadonlyArray<{
        readonly " $fragmentSpreads": FragmentRefs<"MilestonePickerMilestone">;
      } | null | undefined> | null | undefined;
    } | null | undefined;
  } | null | undefined;
};
export type MilestonePickerSearchQuery = {
  response: MilestonePickerSearchQuery$data;
  variables: MilestonePickerSearchQuery$variables;
};